fix: update bottom sheet storybook
